Fix: use connection id to determine if a user is ejected or not
It used to be a boolean value which was durable if the server is
configured to use localStorage for bbb client settings. This would lead
to permanently blocking a user for all future BBB sessions.

To prevent that we use the connectionID to determine if a user is
blocked or not. If a user rejoins (in the same or a different meeting),
the connectionID is different, so the user can join again.