feat(wake-lock): enable implicit activation based on the default setting
This commit removes the wake lock activation toast along with the
enable/disable buttons, implementing the wake lock implicit activation
behavior. The wake lock feature is implicitly activated if the
`defaultSettings.application.wakeLock` in the `settings.yml` file is set
true. For mobile devices that do not support the API or fail when
requesting a wake lock, toasts are raised explaining that calls will be
dropped when the screen turns off.

In cases where `defaultSettings.application.wakeLock` is set false, users
can enable the wake lock manually through the settings menu.