fix(screenshare): use local state as well for rendering
BBB <= 2.7 uses both the local and remote states for loading the
screensharing container in the content area - with proper tracking of
connecting/connected states to speed up UI responsiveness for presenters
and properly reflect the current media state.
That behavior was reverted in 3.0 to only use the remote state, which
regresses the client by causing the UI to be idle (thus not reflect any
state) while screen is being negotiated.

Account for both local and remote states for rendering the screen sharing
component, which fixes the aforementioned regression.