fix(audio): ensure correct audio device labels in Firefox
Firefox incorretly displays placeholder audio device labels in the audio
settings/echo test modal when audio is disconnected. This issue arises
due to two quirks:
- Firefox does not support the 'microphone' query from the Permissions
  API, causing a fallback gUM permission check.
- Firefox omits device labels from `enumerateDevices` if no streams
  are active, even if gUM permission is granted. This behavior differs
  from other browsers and causes our `enumerateDevices` handling to
  assume that granted permission implies labels are present. This
  failed since we clear streams before resolving the fallback gUM.

We now run an additional `enumerateDevices` call in `AudioSettings` when
a selected input device is defined. This ensures `enumerateDevices` is
re-run when a new stream is active, adding the correct device labels in
Firefox and improving device listings in all browsers. We've also
enhanced error handling in the enumeration process and fixed a false
positive in `hasMicrophonePermission`.