docs: change order of mediasoup IP config
See https://github.com/bigbluebutton/bigbluebutton/issues/20966.
The order of the configured mediasoup listenIps influence the server's candidate priorities. 
There's not a big impact to this, but if we want to be friendly to RFC 8421 (and to Happy Eyeballs), we should suggest sorting IPv6 first.

Change docs so that the IPv6 config in mediasoup comes first.