feat(screenshare): add support for troubleshooting links (#21051)
* feat(screenshare): add support for troubleshooting links

Adds setting option to specify troubleshooting links to each error code
of screenshare. When a troubleshooting link for the given error exists,
the toast notification about the error is displayed with a 'Learn more'
button that when clicked leads the user to the external link. When there
is no link set for the specific error code, the button is not displayed.

* fix(screenshare): change toast type for error code 1136

Changed toast type from 'error' to 'warning' for error code 1136 when
sharing screen. This adjustment was made because error code 1136 is also
returned when the user cancels screen sharing during the tab selection
process. Displaying an error toast in this situation could cause
unnecessary alarm for users, as they were simply canceling an operation.

* fix(notification): help link button element

Uses the button element instead of a div to display the 'Learn more'
help link button.

---------

Co-authored-by: Carlos Henrique <carloshsc1998@gmail.com>