fix(layout): focused camera propagation
The focused camera id is of type string and was set to boolean `false`
when a camera was unfocused, causing it to throw an error on apollo and
consequently not propagating the focused camera status. This commit fixes
this issue by setting the focused camera id to its initial string value `none`
when a camera is unfocused.