feat(core): add LiveKit user token generation
LiveKit's access model relies on JWT tokens that need to be generated by
a controlling application in possession of a valid key and secret. We
need a way to generate and bind those to BBB user models.

Implement a pair of RPCs (GenerateWebRtcTokenReq/RespMsg) to be
exchanged between akka-apps and webrtc-sfu, which is BBB's LiveKit
controller. The workflow is for akka-apps to ask for a token to be
generated whenever an user joins, and to listen for token re-generation
whenever webrtc-sfu notifies it.

A new attribute `livekitToken` was added to the user_current DB table so
that client's can pick them up. This is pulled from a remote table
`user_livekit`, which currently only holds the token but will be
expanded in the near future to hold more livekit-user specific info
(e.g.: grant levels and node IDs).

bbb-webrtc-sfu MAY regenerate an user's token token whenever necessary
(e.g.: due to lock settings triggering grant changes). Due to how akka-apps
handle the token generation response, it should work seamlessly.