fix(layout): correct top and height calculation of camera dock bounds
Fixes an issue with the camera dock bounds calculation, where the banner
height was being accounted for twice, leading to incorrect dock height
and excessive spacing. The height is now correctly based on the media
are height, which already factors in the banner height.

Additionally, the top position of the camera dock is now calculated
using the media area's top position, correctly factoring in the margin of
the media area.