build: Umask fixes for 2.7 build scripts (#21366)
* Fix permissions during build of bbb-libreoffice-docker

The scripts must be executable by anyone. They are owned by root and
executed by user bigbluebutton.

* Fix permissions during build of bbb-html5

The files must be readable by mongodb. There are no secrets in the file,
so it is safe to be readable by anyone.

* Fix permissions during build of bbb-record-core

Files must be readable by user bigbluebutton.

* Fix more umask issues

Recording scripts were not executable in the builds

---------

Co-authored-by: Daniel Schreiber <daniel.schreiber@hrz.tu-chemnitz.de>