fix(audio): retry gUM on OverconstrainedError
Whenever echo is skipped, audio's gUM routine is _not_ retrying with
rolled back constraints when it fails due to an OverconstrainedError.
This ends up generating an error for the end user while it should just
try to connect to audio with fresh constraints instead.

Use doGUM's retryOnFailure flag to properly retry gUM with rolled back
constraints when necessary.