fix(audio): make WS errors retryable
SFU's audio bridge only retries audio when join fails if
RETRY_THROUGH_RELAY is active. This ignores 1010/1007s when
RETRY_THROUGH_RELAY is false or WS establishment errors happens, which
is not ideal as both cases are worth a retry.

Adjust the retry procedures so that both non-relay 1007/1010 and
1002/1005 (WS) errors cause a retry.
This commit also fixes an issue with the error modal not being shown
correctly when a retry attempt failed - the reconnect method was not
relaying the result of the retry attempt to the original joinAudio
Promise.

The logCode `sfuaudio_error_retry_through_relay` was adjusted to
`sfuaudio_error_retry` - and forceRelay was added to its extraInfo - so
that it correctly depicts all retry scenarios.