refactor(audio): make audio join cancel configurable
Audio was made cancellable a while ago by allowing users to close the
audio modal while audio is being negotiated. By looking through metrics,
I have a theory that this may be 1) increasing error rates artificially
2) confusing end users who mistakengly click outside of the modal.

Cancelling audio join should not be an option in the first place if
there's no good way to let users know how to roll back from that
situation (there isn't, right now). There are also timeouts in place in
both audio bridges so that it doesn't stay "loading" forever.

This commit adds a new flag `public.media.audio.allowAudioJoinCancel`
that makes the aforementioned behavior configurable. Default value is
preserved (true, allow). The idea is for this flag to be used in field
trials to validate behaviors 1) and 2) - and then decide what's the best
UX flow.