fix(audio): notify user that a join retry is happening
When joining audio fails and a retry attempt is underway, there's no
indication in the UI (i.e.: in AudioModal's "Establishing audio
connection..." screen) that it's happening. Since the whole cycle may
take a while (up to 40-50 seconds with retries), letting the end user
know that something is _happening_, even if it's taking longer than
expected, would be good.

Add a subtext to AudioModal's audio establishment view whenever a retry
is underway (i.e.: isReconnecting=true). The subtext is intentionally
not technical (`This is taking longer than expected. Please wait a
moment.`) as the goal is just to provide users with an update.