refactor(audio): use simplest constraints possible on gUM rollback
There's no hard indications that this is a problem, but some weird user
logs after a gUM rollback makes me thing we should err on the side of
caution and use the simplesment constraints possible when rolling back.

Modify the fallback constraints to be a simple { audio: true } - hence
ignoring not just the supplied deviceId, but any filter options that
were inherited by config (server or client).